requirements.txt
pandas
numpy
requests
snowflake-connector-python
python-dotenv
ipykernel
pyarrow
matplotlib
seaborn
scikit-learn
statsmodels
xgboost
category_encoders
catboost
mlflow
prophet
pmdarima
streamlit
streamlit-shap



